Ugh, broken pipe.
2001-10-27  Anders Carlsson  <andersca@gnu.org>

	* gdk/x11/gdkmain-x11.c: Replace int and guint with
	long and gulong in some places to make the code 64-bit
	clean.
	(_gdk_windowing_init_check): Replace gint with gulong.
	This patch was made by George Lebl and fixes bug #62113.

+++ gdk/x11/gdkmain-x11.c
@@ -115,7 +115,7 @@ _gdk_windowing_init_check (int argc, char **argv)
 {
   XKeyboardState keyboard_state;
   XClassHint *class_hint;
-  guint pid;
+  gulong pid;
 
   _gdk_x11_initialize_locale ();
